organize code, TODO create argparse functionality in main
