+++ src/__init__.py
@@ -1,0 +1,5 @@
+"""
+Package enabling computation accross the different Pokemon types.
+"""
+from pktype import PKType as PKType
+from combined_pktype import CombinedPKType as CombinedPKType
